refactor: organize agentboot public models
